fix: ignore demo 404/403s in Execute view smoke test
Keep Execute Playwright helpers out of shared helpers.js so CI does not
run every module, including catalyst-showcase which has no tests.
